+++ packages/cloud_functions/cloud_functions_web/lib/utils.dart
@@ -8,7 +8,7 @@ import 'package:cloud_functions_platform_interface/cloud_functions_platform_inte
 import 'package:firebase_core_web/firebase_core_web_interop.dart' show dartify;
 
 /// Given a web error, a [FirebaseFunctionsException] is returned.
-FirebaseFunctionsException throwFirebaseFunctionsException(
+FirebaseFunctionsException convertFirebaseFunctionsException(
   Object exception, [
   StackTrace? stackTrace,
 ]) {