+++ packages/firebase_auth/firebase_auth/android/src/main/java/io/flutter/plugins/firebase/auth/Constants.java
@@ -56,7 +56,6 @@ public class Constants {
   public static final String ACCESS_TOKEN = "accessToken";
   public static final String CODE = "code";
   public static final String RAW_NONCE = "rawNonce";
-  public static final String HANDLE = "handle";
   public static final String EMAIL_LINK = "emailLink";
   public static final String VERIFICATION_ID = "verificationId";
   public static final String SMS_CODE = "smsCode";
@@ -79,4 +78,5 @@ public class Constants {
   public static final String AUTO_RETRIEVED_SMS_CODE_FOR_TESTING = "autoRetrievedSmsCodeForTesting";
   public static final String HOST = "host";
   public static final String PORT = "port";
+  public static final String NAME = "name";
 }