+++ packages/firebase_auth/firebase_auth/android/src/main/java/io/flutter/plugins/firebase/auth/FlutterFirebaseAuthPlugin.java
@@ -8,14 +8,12 @@ import static io.flutter.plugins.firebase.core.FlutterFirebasePluginRegistry.reg
 
 import android.app.Activity;
 import android.net.Uri;
-import android.util.Log;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.Tasks;
 import com.google.firebase.FirebaseApiNotAvailableException;
 import com.google.firebase.FirebaseApp;
-import com.google.firebase.FirebaseException;
 import com.google.firebase.FirebaseNetworkException;
 import com.google.firebase.FirebaseTooManyRequestsException;
 import com.google.firebase.auth.ActionCodeEmailInfo;
@@ -37,7 +35,6 @@ import com.google.firebase.auth.GithubAuthProvider;
 import com.google.firebase.auth.GoogleAuthProvider;
 import com.google.firebase.auth.OAuthProvider;
 import com.google.firebase.auth.PhoneAuthCredential;
-import com.google.firebase.auth.PhoneAuthOptions;
 import com.google.firebase.auth.PhoneAuthProvider;
 import com.google.firebase.auth.SignInMethodQueryResult;
 import com.google.firebase.auth.TwitterAuthProvider;
@@ -47,6 +44,8 @@ import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.embedding.engine.plugins.activity.ActivityAware;
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
 import io.flutter.plugin.common.BinaryMessenger;
+import io.flutter.plugin.common.EventChannel;
+import io.flutter.plugin.common.EventChannel.StreamHandler;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
@@ -55,35 +54,31 @@ import io.flutter.plugin.common.PluginRegistry;
 import io.flutter.plugins.firebase.core.FlutterFirebasePlugin;
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.UUID;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
 
 /** Flutter plugin for Firebase Auth. */
 public class FlutterFirebaseAuthPlugin
     implements FlutterFirebasePlugin, MethodCallHandler, FlutterPlugin, ActivityAware {
 
+  private static final String METHOD_CHANNEL_NAME = "plugins.flutter.io/firebase_auth";
+
   // Stores the instances of native AuthCredentials by their hashCode
   static final HashMap<Integer, AuthCredential> authCredentials = new HashMap<>();
 
-  private static final HashMap<String, FirebaseAuth.AuthStateListener> authListeners =
-      new HashMap<>();
-  private static final HashMap<String, FirebaseAuth.IdTokenListener> idTokenListeners =
-      new HashMap<>();
-  private static final HashMap<Integer, PhoneAuthProvider.ForceResendingToken>
-      forceResendingTokens = new HashMap<>();
-  private PluginRegistry.Registrar registrar;
+  @Nullable private BinaryMessenger messenger;
+
   private MethodChannel channel;
   private Activity activity;
-  private static Boolean initialAuthState = true;
+
+  private final Map<EventChannel, StreamHandler> streamHandlers = new HashMap<>();
 
   @SuppressWarnings("unused")
   public static void registerWith(PluginRegistry.Registrar registrar) {
     FlutterFirebaseAuthPlugin instance = new FlutterFirebaseAuthPlugin();
-    instance.registrar = registrar;
     instance.initInstance(registrar.messenger());
   }
 
@@ -105,10 +100,11 @@ public class FlutterFirebaseAuthPlugin
   }
 
   private void initInstance(BinaryMessenger messenger) {
-    String channelName = "plugins.flutter.io/firebase_auth";
-    registerPlugin(channelName, this);
-    channel = new MethodChannel(messenger, channelName);
+    registerPlugin(METHOD_CHANNEL_NAME, this);
+    channel = new MethodChannel(messenger, METHOD_CHANNEL_NAME);
     channel.setMethodCallHandler(this);
+
+    this.messenger = messenger;
   }
 
   @Override
@@ -118,9 +114,11 @@ public class FlutterFirebaseAuthPlugin
 
   @Override
   public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
-    removeEventListeners();
     channel.setMethodCallHandler(null);
     channel = null;
+    messenger = null;
+
+    removeEventListeners();
   }
 
   @Override
@@ -143,64 +141,16 @@ public class FlutterFirebaseAuthPlugin
     activity = null;
   }
 
-  // Ensure any listeners are removed when the app
-  // is detached from the FlutterEngine
-  private void removeEventListeners() {
-    Iterator<?> authListenerIterator = authListeners.entrySet().iterator();
-
-    while (authListenerIterator.hasNext()) {
-      Map.Entry<?, ?> pair = (Map.Entry<?, ?>) authListenerIterator.next();
-      String appName = (String) pair.getKey();
-      FirebaseApp firebaseApp = FirebaseApp.getInstance(appName);
-      FirebaseAuth firebaseAuth = FirebaseAuth.getInstance(firebaseApp);
-      FirebaseAuth.AuthStateListener authListener =
-          (FirebaseAuth.AuthStateListener) pair.getValue();
-      firebaseAuth.removeAuthStateListener(authListener);
-      authListenerIterator.remove();
-    }
-
-    Iterator<?> idTokenListenerIterator = idTokenListeners.entrySet().iterator();
-
-    while (idTokenListenerIterator.hasNext()) {
-      Map.Entry<?, ?> pair = (Map.Entry<?, ?>) idTokenListenerIterator.next();
-      String appName = (String) pair.getKey();
-      FirebaseApp firebaseApp = FirebaseApp.getInstance(appName);
-      FirebaseAuth firebaseAuth = FirebaseAuth.getInstance(firebaseApp);
-      FirebaseAuth.IdTokenListener authListener = (FirebaseAuth.IdTokenListener) pair.getValue();
-      firebaseAuth.removeIdTokenListener(authListener);
-      idTokenListenerIterator.remove();
-    }
-  }
-
   // Only access activity with this method.
+  @Nullable
   private Activity getActivity() {
-    return registrar != null ? registrar.activity() : activity;
+    return activity;
   }
 
-  private FirebaseAuth getAuth(Map<String, Object> arguments) {
+  static FirebaseAuth getAuth(Map<String, Object> arguments) {
     String appName = (String) Objects.requireNonNull(arguments.get(Constants.APP_NAME));
     FirebaseApp app = FirebaseApp.getInstance(appName);
     return FirebaseAuth.getInstance(app);
-  }
-
-  private MethodChannel.Result getMethodChannelResultHandler(String method) {
-    return new Result() {
-      @Override
-      public void success(@Nullable Object result) {
-        // Noop
-      }
-
-      @Override
-      public void error(
-          String errorCode, @Nullable String errorMessage, @Nullable Object errorDetails) {
-        Log.e(Constants.TAG, method + " error (" + errorCode + "): " + errorMessage);
-      }
-
-      @Override
-      public void notImplemented() {
-        Log.e(Constants.TAG, method + " has not been implemented");
-      }
-    };
   }
 
   private FirebaseUser getCurrentUser(Map<String, Object> arguments) {
@@ -360,7 +310,7 @@ public class FlutterFirebaseAuthPlugin
   }
 
   @SuppressWarnings("ConstantConditions")
-  private Map<String, Object> parseFirebaseUser(FirebaseUser firebaseUser) {
+  static Map<String, Object> parseFirebaseUser(FirebaseUser firebaseUser) {
     if (firebaseUser == null) {
       return null;
     }
@@ -391,7 +341,8 @@ public class FlutterFirebaseAuthPlugin
     return output;
   }
 
-  private List<Map<String, Object>> parseUserInfoList(List<? extends UserInfo> userInfoList) {
+  private static List<Map<String, Object>> parseUserInfoList(
+      List<? extends UserInfo> userInfoList) {
     List<Map<String, Object>> output = new ArrayList<>();
 
     for (UserInfo userInfo : userInfoList) {
@@ -404,7 +355,7 @@ public class FlutterFirebaseAuthPlugin
   }
 
   @SuppressWarnings("ConstantConditions")
-  private Map<String, Object> parseUserInfo(@NonNull UserInfo userInfo) {
+  private static Map<String, Object> parseUserInfo(@NonNull UserInfo userInfo) {
     Map<String, Object> output = new HashMap<>();
 
     output.put(Constants.DISPLAY_NAME, userInfo.getDisplayName());
@@ -417,7 +368,7 @@ public class FlutterFirebaseAuthPlugin
     return output;
   }
 
-  private String parsePhotoUrl(Uri photoUri) {
+  private static String parsePhotoUrl(Uri photoUri) {
     if (photoUri == null) {
       return null;
     }
@@ -493,67 +444,31 @@ public class FlutterFirebaseAuthPlugin
     return output;
   }
 
-  @SuppressWarnings("ConstantConditions")
-  private Task<Void> registerChangeListeners(Map<String, Object> arguments) {
+  private Task<String> registerIdTokenListener(Map<String, Object> arguments) {
     return Tasks.call(
         cachedThreadPool,
         () -> {
-          String appName = (String) Objects.requireNonNull(arguments.get(Constants.APP_NAME));
-          FirebaseAuth firebaseAuth = getAuth(arguments);
+          final FirebaseAuth auth = getAuth(arguments);
+          final IdTokenChannelStreamHandler handler = new IdTokenChannelStreamHandler(auth);
+          final String name = METHOD_CHANNEL_NAME + "/id-token/" + auth.getApp().getName();
+          final EventChannel channel = new EventChannel(messenger, name);
+          channel.setStreamHandler(handler);
+          streamHandlers.put(channel, handler);
+          return name;
+        });
+  }
 
-          FirebaseAuth.AuthStateListener authStateListener = authListeners.get(appName);
-          FirebaseAuth.IdTokenListener idTokenListener = idTokenListeners.get(appName);
-
-          Map<String, Object> event = new HashMap<>();
-          event.put(Constants.APP_NAME, appName);
-
-          if (authStateListener == null) {
-            FirebaseAuth.AuthStateListener newAuthStateListener =
-                auth -> {
-                  FirebaseUser user = auth.getCurrentUser();
-
-                  if (user == null) {
-                    event.put(Constants.USER, null);
-                  } else {
-                    event.put(Constants.USER, parseFirebaseUser(user));
-                  }
-
-                  if (initialAuthState) {
-                    initialAuthState = false;
-                  } else {
-                    channel.invokeMethod(
-                        "Auth#authStateChanges",
-                        event,
-                        getMethodChannelResultHandler("Auth#authStateChanges"));
-                  }
-                };
-
-            firebaseAuth.addAuthStateListener(newAuthStateListener);
-            authListeners.put(appName, newAuthStateListener);
-          }
-
-          if (idTokenListener == null) {
-            FirebaseAuth.IdTokenListener newIdTokenChangeListener =
-                auth -> {
-                  FirebaseUser user = auth.getCurrentUser();
-
-                  if (user == null) {
-                    event.put(Constants.USER, null);
-                  } else {
-                    event.put(Constants.USER, parseFirebaseUser(user));
-                  }
-
-                  channel.invokeMethod(
-                      "Auth#idTokenChanges",
-                      event,
-                      getMethodChannelResultHandler("Auth#idTokenChanges"));
-                };
-
-            firebaseAuth.addIdTokenListener(newIdTokenChangeListener);
-            idTokenListeners.put(appName, newIdTokenChangeListener);
-          }
-
-          return null;
+  private Task<String> registerAuthStateListener(Map<String, Object> arguments) {
+    return Tasks.call(
+        cachedThreadPool,
+        () -> {
+          final FirebaseAuth auth = getAuth(arguments);
+          final AuthStateChannelStreamHandler handler = new AuthStateChannelStreamHandler(auth);
+          final String name = METHOD_CHANNEL_NAME + "/auth-state/" + auth.getApp().getName();
+          final EventChannel channel = new EventChannel(messenger, name);
+          channel.setStreamHandler(handler);
+          streamHandlers.put(channel, handler);
+          return name;
         });
   }
 
@@ -789,111 +704,25 @@ public class FlutterFirebaseAuthPlugin
         });
   }
 
-  private Task<Void> verifyPhoneNumber(Map<String, Object> arguments) {
+  private Task<String> verifyPhoneNumber(Map<String, Object> arguments) {
     return Tasks.call(
         cachedThreadPool,
         () -> {
-          FirebaseAuth firebaseAuth = getAuth(arguments);
-          String phoneNumber =
-              (String) Objects.requireNonNull(arguments.get(Constants.PHONE_NUMBER));
-          int handle = (int) Objects.requireNonNull(arguments.get(Constants.HANDLE));
-          int timeout = (int) Objects.requireNonNull(arguments.get(Constants.TIMEOUT));
+          String eventChannelName = METHOD_CHANNEL_NAME + "/phone/" + UUID.randomUUID().toString();
+          EventChannel channel = new EventChannel(messenger, eventChannelName);
+          PhoneNumberVerificationStreamHandler handler =
+              new PhoneNumberVerificationStreamHandler(
+                  getActivity(),
+                  arguments,
+                  credential -> {
+                    int hashCode = credential.hashCode();
+                    authCredentials.put(hashCode, credential);
+                  });
 
-          Map<String, Object> event = new HashMap<>();
-          event.put(Constants.HANDLE, handle);
+          channel.setStreamHandler(handler);
+          streamHandlers.put(channel, handler);
 
-          PhoneAuthProvider.OnVerificationStateChangedCallbacks callbacks =
-              new PhoneAuthProvider.OnVerificationStateChangedCallbacks() {
-                @Override
-                public void onVerificationCompleted(
-                    @NonNull PhoneAuthCredential phoneAuthCredential) {
-                  int phoneAuthCredentialHashCode = phoneAuthCredential.hashCode();
-                  authCredentials.put(phoneAuthCredentialHashCode, phoneAuthCredential);
-                  event.put(Constants.TOKEN, phoneAuthCredentialHashCode);
-
-                  if (phoneAuthCredential.getSmsCode() != null) {
-                    event.put(Constants.SMS_CODE, phoneAuthCredential.getSmsCode());
-                  }
-
-                  channel.invokeMethod(
-                      "Auth#phoneVerificationCompleted",
-                      event,
-                      getMethodChannelResultHandler("Auth#phoneVerificationCompleted"));
-                }
-
-                @Override
-                public void onVerificationFailed(@NonNull FirebaseException e) {
-                  Map<String, Object> error = new HashMap<>();
-                  error.put("message", e.getLocalizedMessage());
-                  error.put("details", getExceptionDetails(e));
-                  event.put("error", error);
-
-                  channel.invokeMethod(
-                      "Auth#phoneVerificationFailed",
-                      event,
-                      getMethodChannelResultHandler("Auth#phoneVerificationFailed"));
-                }
-
-                @Override
-                public void onCodeSent(
-                    @NonNull String verificationId,
-                    @NonNull PhoneAuthProvider.ForceResendingToken token) {
-                  int forceResendingTokenHashCode = token.hashCode();
-                  forceResendingTokens.put(forceResendingTokenHashCode, token);
-                  event.put(Constants.VERIFICATION_ID, verificationId);
-                  event.put(Constants.FORCE_RESENDING_TOKEN, forceResendingTokenHashCode);
-
-                  channel.invokeMethod(
-                      "Auth#phoneCodeSent",
-                      event,
-                      getMethodChannelResultHandler("Auth#phoneCodeSent"));
-                }
-
-                @Override
-                public void onCodeAutoRetrievalTimeOut(@NonNull String verificationId) {
-                  event.put(Constants.VERIFICATION_ID, verificationId);
-
-                  channel.invokeMethod(
-                      "Auth#phoneCodeAutoRetrievalTimeout",
-                      event,
-                      getMethodChannelResultHandler("Auth#phoneCodeAutoRetrievalTimeout"));
-                }
-              };
-
-          // Allows the auto-retrieval flow to be tested.
-          // See https://firebase.google.com/docs/auth/android/phone-auth#integration-testing
-          if (arguments.get(Constants.AUTO_RETRIEVED_SMS_CODE_FOR_TESTING) != null) {
-            String autoRetrievedSmsCodeForTesting =
-                (String)
-                    Objects.requireNonNull(
-                        arguments.get(Constants.AUTO_RETRIEVED_SMS_CODE_FOR_TESTING));
-
-            firebaseAuth
-                .getFirebaseAuthSettings()
-                .setAutoRetrievedSmsCodeForPhoneNumber(phoneNumber, autoRetrievedSmsCodeForTesting);
-          }
-
-          PhoneAuthOptions.Builder phoneAuthOptionsBuilder =
-              new PhoneAuthOptions.Builder(firebaseAuth);
-          phoneAuthOptionsBuilder.setActivity(getActivity());
-          phoneAuthOptionsBuilder.setPhoneNumber(phoneNumber);
-          phoneAuthOptionsBuilder.setCallbacks(callbacks);
-          phoneAuthOptionsBuilder.setTimeout((long) timeout, TimeUnit.MILLISECONDS);
-
-          if (arguments.get(Constants.FORCE_RESENDING_TOKEN) != null) {
-            int forceResendingTokenHashCode =
-                (int) Objects.requireNonNull(arguments.get(Constants.FORCE_RESENDING_TOKEN));
-
-            PhoneAuthProvider.ForceResendingToken forceResendingToken =
-                forceResendingTokens.get(forceResendingTokenHashCode);
-
-            if (forceResendingToken != null) {
-              phoneAuthOptionsBuilder.setForceResendingToken(forceResendingToken);
-            }
-          }
-
-          PhoneAuthProvider.verifyPhoneNumber(phoneAuthOptionsBuilder.build());
-          return null;
+          return eventChannelName;
         });
   }
 
@@ -1157,11 +986,14 @@ public class FlutterFirebaseAuthPlugin
 
   @Override
   public void onMethodCall(@NonNull MethodCall call, @NonNull Result result) {
-    Task<?> methodCallTask;
+    final Task<?> methodCallTask;
 
     switch (call.method) {
-      case "Auth#registerChangeListeners":
-        methodCallTask = registerChangeListeners(call.arguments());
+      case "Auth#registerIdTokenListener":
+        methodCallTask = registerIdTokenListener(call.arguments());
+        break;
+      case "Auth#registerAuthStateListener":
+        methodCallTask = registerAuthStateListener(call.arguments());
         break;
       case "Auth#applyActionCode":
         methodCallTask = applyActionCode(call.arguments());
@@ -1296,7 +1128,7 @@ public class FlutterFirebaseAuthPlugin
         });
   }
 
-  private Map<String, Object> getExceptionDetails(Exception exception) {
+  static Map<String, Object> getExceptionDetails(Exception exception) {
     Map<String, Object> details = new HashMap<>();
 
     if (exception == null) {
@@ -1379,8 +1211,16 @@ public class FlutterFirebaseAuthPlugin
         () -> {
           removeEventListeners();
           authCredentials.clear();
-          forceResendingTokens.clear();
           return null;
         });
+  }
+
+  private void removeEventListeners() {
+    for (EventChannel eventChannel : streamHandlers.keySet()) {
+      StreamHandler streamHandler = streamHandlers.get(eventChannel);
+      streamHandler.onCancel(null);
+      eventChannel.setStreamHandler(null);
+    }
+    streamHandlers.clear();
   }
 }