+++ packages/firebase_auth/firebase_auth/ios/Classes/Public/FLTFirebaseAuthPlugin.h
@@ -12,4 +12,9 @@
 #import <firebase_core/FLTFirebasePlugin.h>
 
 @interface FLTFirebaseAuthPlugin : FLTFirebasePlugin <FlutterPlugin, FLTFirebasePlugin>
+
++ (id)getNSDictionaryFromAuthCredential:(FIRAuthCredential *)authCredential;
++ (NSDictionary *)getNSDictionaryFromUserInfo:(id<FIRUserInfo>)userInfo;
++ (NSMutableDictionary *)getNSDictionaryFromUser:(FIRUser *)user;
++ (NSDictionary *)getNSDictionaryFromNSError:(NSError *)error;
 @end