+++ packages/firebase_auth/firebase_auth/test/firebase_auth_test.dart
@@ -428,10 +428,11 @@ void main() {
         final AuthCredential credential =
             GithubAuthProvider.credential(kMockGithubToken);
         await auth!.signInWithCredential(credential);
-        final GithubAuthCredential captured =
+        final captured =
             verify(mockAuthPlatform.signInWithCredential(captureAny))
                 .captured
                 .single;
+        expect(captured, isA<GithubAuthCredential>());
         expect(captured.providerId, equals('github.com'));
         expect(captured.accessToken, equals(kMockGithubToken));
       });
@@ -458,25 +459,27 @@ void main() {
           secret: kMockAccessToken,
         );
         await auth!.signInWithCredential(credential);
-        final TwitterAuthCredential captured =
+        final captured =
             verify(mockAuthPlatform.signInWithCredential(captureAny))
                 .captured
                 .single;
+        expect(captured, isA<TwitterAuthCredential>());
         expect(captured.providerId, equals('twitter.com'));
         expect(captured.accessToken, equals(kMockIdToken));
         expect(captured.secret, equals(kMockAccessToken));
       });
 
       test('GoogleAuthProvider signInWithCredential', () async {
-        final AuthCredential credential = GoogleAuthProvider.credential(
+        final credential = GoogleAuthProvider.credential(
           idToken: kMockIdToken,
           accessToken: kMockAccessToken,
         );
         await auth!.signInWithCredential(credential);
-        final GoogleAuthCredential captured =
+        final captured =
             verify(mockAuthPlatform.signInWithCredential(captureAny))
                 .captured
                 .single;
+        expect(captured, isA<GoogleAuthCredential>());
         expect(captured.providerId, equals('google.com'));
         expect(captured.idToken, equals(kMockIdToken));
         expect(captured.accessToken, equals(kMockAccessToken));
@@ -489,7 +492,7 @@ void main() {
           accessToken: kMockAccessToken,
         );
         await auth!.signInWithCredential(credential);
-        final OAuthCredential captured =
+        final captured =
             verify(mockAuthPlatform.signInWithCredential(captureAny))
                 .captured
                 .single;
@@ -517,10 +520,11 @@ void main() {
         final AuthCredential credential =
             FacebookAuthProvider.credential(kMockAccessToken);
         await auth!.signInWithCredential(credential);
-        final FacebookAuthCredential captured =
+        final captured =
             verify(mockAuthPlatform.signInWithCredential(captureAny))
                 .captured
                 .single;
+        expect(captured, isA<FacebookAuthCredential>());
         expect(captured.providerId, equals('facebook.com'));
         expect(captured.accessToken, equals(kMockAccessToken));
       });