+++ packages/firebase_auth/firebase_auth_platform_interface/lib/src/method_channel/method_channel_firebase_auth.dart
@@ -4,31 +4,20 @@
 
 import 'dart:async';
 
-import '../../firebase_auth_platform_interface.dart';
-import '../firebase_auth_exception.dart';
-import 'method_channel_user.dart';
-import '../platform_interface/platform_interface_user_credential.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
+import '../../firebase_auth_platform_interface.dart';
+import '../firebase_auth_exception.dart';
+import '../platform_interface/platform_interface_user_credential.dart';
+import 'method_channel_user.dart';
 import 'method_channel_user_credential.dart';
 import 'utils/exception.dart';
-import 'utils/phone_auth_callbacks.dart';
 
 /// Method Channel delegate for [FirebaseAuthPlatform].
 class MethodChannelFirebaseAuth extends FirebaseAuthPlatform {
-  /// Keep an internal reference to whether the [MethodChannelFirebaseAuth]
-  /// class has already been initialized.
-  static bool _initialized = false;
-
-  /// Keeps an internal handle ID for the channel.
-  static int _methodChannelHandleId = 0;
-
-  /// Increments and returns the next channel ID handler for Auth.
-  static int get nextMethodChannelHandleId => _methodChannelHandleId++;
-
   /// The [MethodChannelFirebaseAuth] method channel.
   static const MethodChannel channel = MethodChannel(
     'plugins.flutter.io/firebase_auth',
@@ -38,16 +27,21 @@ class MethodChannelFirebaseAuth extends FirebaseAuthPlatform {
       _methodChannelFirebaseAuthInstances =
       <String, MethodChannelFirebaseAuth>{};
 
-  static Map<String, StreamController<UserPlatform?>>
-      _authStateChangesListeners = <String, StreamController<UserPlatform?>>{};
+  static final Map<String, StreamController<_ValueWrapper<UserPlatform>>>
+      _authStateChangesListeners =
+      <String, StreamController<_ValueWrapper<UserPlatform>>>{};
 
-  static Map<String, StreamController<UserPlatform?>> _idTokenChangesListeners =
-      <String, StreamController<UserPlatform?>>{};
+  static final Map<String, StreamController<_ValueWrapper<UserPlatform>>>
+      _idTokenChangesListeners =
+      <String, StreamController<_ValueWrapper<UserPlatform>>>{};
 
-  static Map<String, StreamController<UserPlatform?>> _userChangesListeners =
-      <String, StreamController<UserPlatform?>>{};
+  static final Map<String, StreamController<_ValueWrapper<UserPlatform>>>
+      _userChangesListeners =
+      <String, StreamController<_ValueWrapper<UserPlatform>>>{};
 
-  static Map<int, PhoneAuthCallbacks> _phoneAuthCallbacks = {};
+  StreamController<T> _createBroadcastStream<T>() {
+    return StreamController<T>.broadcast();
+  }
 
   /// Returns a stub instance to allow the platform interface to access
   /// the class instance statically.
@@ -64,47 +58,35 @@ class MethodChannelFirebaseAuth extends FirebaseAuthPlatform {
   /// Creates a new instance with a given [FirebaseApp].
   MethodChannelFirebaseAuth({required FirebaseApp app})
       : super(appInstance: app) {
-    // Send a request to start listening to change listeners straight away
-    channel
-        .invokeMethod<void>('Auth#registerChangeListeners', <String, dynamic>{
+    channel.invokeMethod<String>('Auth#registerIdTokenListener', {
       'appName': app.name,
+    }).then((channelName) {
+      final events = EventChannel(channelName!, channel.codec);
+      events.receiveBroadcastStream().listen(
+        (arguments) {
+          _handleIdTokenChangesListener(app.name, arguments);
+        },
+      );
+    });
+
+    channel.invokeMethod<String>('Auth#registerAuthStateListener', {
+      'appName': app.name,
+    }).then((channelName) {
+      final events = EventChannel(channelName!, channel.codec);
+      events.receiveBroadcastStream().listen(
+        (arguments) {
+          _handleAuthStateChangesListener(app.name, arguments);
+        },
+      );
     });
 
     // Create a app instance broadcast stream for native listener events
     _authStateChangesListeners[app.name] =
-        StreamController<UserPlatform?>.broadcast();
+        _createBroadcastStream<_ValueWrapper<UserPlatform>>();
     _idTokenChangesListeners[app.name] =
-        StreamController<UserPlatform?>.broadcast();
+        _createBroadcastStream<_ValueWrapper<UserPlatform>>();
     _userChangesListeners[app.name] =
-        StreamController<UserPlatform?>.broadcast();
-
-    // The channel setMethodCallHandler callback is not app specific, so there
-    // is no need to register the caller more than once.
-    if (_initialized) return;
-
-    channel.setMethodCallHandler((MethodCall call) async {
-      Map<dynamic, dynamic> arguments = call.arguments as Map<dynamic, dynamic>;
-
-      switch (call.method) {
-        case 'Auth#authStateChanges':
-          // TODO(rrousselGit): find out why an exception here isn't reported to tests
-          return _handleAuthStateChangesListener(arguments);
-        case 'Auth#idTokenChanges':
-          return _handleIdTokenChangesListener(arguments);
-        case 'Auth#phoneVerificationCompleted':
-          return _handlePhoneVerificationCompleted(arguments);
-        case 'Auth#phoneVerificationFailed':
-          return _handlePhoneVerificationFailed(arguments);
-        case 'Auth#phoneCodeSent':
-          return _handlePhoneCodeSent(arguments);
-        case 'Auth#phoneCodeAutoRetrievalTimeout':
-          return _handlePhoneCodeAutoRetrievalTimeout(arguments);
-        default:
-          throw UnimplementedError('${call.method} has not been implemented');
-      }
-    });
-
-    _initialized = true;
+        _createBroadcastStream<_ValueWrapper<UserPlatform>>();
   }
 
   @override
@@ -117,34 +99,30 @@ class MethodChannelFirebaseAuth extends FirebaseAuthPlatform {
   void sendAuthChangesEvent(String appName, UserPlatform? userPlatform) {
     assert(_userChangesListeners[appName] != null);
 
-    _userChangesListeners[appName]!.add(userPlatform);
+    _userChangesListeners[appName]!.add(_ValueWrapper(userPlatform));
   }
 
   /// Handles any incoming [authChanges] listener events.
   // Duplicate setting of [currentUser] in [_handleAuthStateChangesListener] & [_handleIdTokenChangesListener]
   // as iOS & Android do not guarantee correct ordering
   Future<void> _handleAuthStateChangesListener(
-    Map<dynamic, dynamic> arguments,
-  ) async {
-    String appName = arguments['appName'];
-    // ignore: close_sinks, the sink is not created here, only obtained
-    StreamController<UserPlatform?> streamController =
-        _authStateChangesListeners[appName]!;
+      String appName, Map<dynamic, dynamic> arguments) async {
+    // ignore: close_sinks
+    final streamController = _authStateChangesListeners[appName]!;
     MethodChannelFirebaseAuth instance =
         _methodChannelFirebaseAuthInstances[appName]!;
 
-    final Map? userMap = arguments['user'] as Map?;
-
+    final userMap = arguments['user'];
     if (userMap == null) {
       instance.currentUser = null;
-      streamController.add(null);
+      streamController.add(const _ValueWrapper.absent());
     } else {
       final MethodChannelUser user =
           MethodChannelUser(instance, userMap.cast<String, dynamic>());
 
       // TODO(rousselGit): should this logic be moved to the setter instead?
       instance.currentUser = user;
-      streamController.add(instance.currentUser);
+      streamController.add(_ValueWrapper(instance.currentUser));
     }
   }
 
@@ -153,87 +131,30 @@ class MethodChannelFirebaseAuth extends FirebaseAuthPlatform {
   /// This handler also manages the [currentUser] along with sending events
   /// to any [userChanges] stream subscribers.
   Future<void> _handleIdTokenChangesListener(
-    Map<dynamic, dynamic> arguments,
-  ) async {
-    String appName = arguments['appName'];
-    // ignore: close_sinks, the sink is not created here, only obtained
-    StreamController<UserPlatform?> idTokenStreamController =
-        _idTokenChangesListeners[appName]!;
-    // ignore: close_sinks, the sink is not created here, only obtained
-    StreamController<UserPlatform?> userChangesStreamController =
-        _userChangesListeners[appName]!;
+      String appName, Map<dynamic, dynamic> arguments) async {
+    final StreamController<_ValueWrapper<UserPlatform>>
+        // ignore: close_sinks
+        idTokenStreamController = _idTokenChangesListeners[appName]!;
+    final StreamController<_ValueWrapper<UserPlatform>>
+        // ignore: close_sinks
+        userChangesStreamController = _userChangesListeners[appName]!;
     MethodChannelFirebaseAuth instance =
         _methodChannelFirebaseAuthInstances[appName]!;
 
-    final Map? userMap = arguments['user'] as Map?;
-
+    final userMap = arguments['user'];
     if (userMap == null) {
       instance.currentUser = null;
-      idTokenStreamController.add(null);
-      userChangesStreamController.add(null);
+      idTokenStreamController.add(const _ValueWrapper.absent());
+      userChangesStreamController.add(const _ValueWrapper.absent());
     } else {
       final MethodChannelUser user =
           MethodChannelUser(instance, userMap.cast<String, dynamic>());
 
       // TODO(rousselGit): should this logic be moved to the setter instead?
       instance.currentUser = user;
-      idTokenStreamController.add(user);
-      userChangesStreamController.add(user);
+      idTokenStreamController.add(_ValueWrapper(user));
+      userChangesStreamController.add(_ValueWrapper(user));
     }
-  }
-
-  Future<void> _handlePhoneVerificationCompleted(
-    Map<dynamic, dynamic> arguments,
-  ) async {
-    final int handle = arguments['handle'];
-    final int token = arguments['token'];
-    final String smsCode = arguments['smsCode'];
-
-    PhoneAuthCredential phoneAuthCredential =
-        PhoneAuthProvider.credentialFromToken(
-      token,
-      smsCode: smsCode,
-    );
-    PhoneAuthCallbacks callbacks = _phoneAuthCallbacks[handle]!;
-    callbacks.verificationCompleted(phoneAuthCredential);
-  }
-
-  Future<void> _handlePhoneVerificationFailed(
-    Map<dynamic, dynamic> arguments,
-  ) async {
-    final int handle = arguments['handle'];
-    final Map<dynamic, dynamic> error = arguments['error'];
-    final Map<dynamic, dynamic>? details = error['details'];
-
-    PhoneAuthCallbacks callbacks = _phoneAuthCallbacks[handle]!;
-
-    FirebaseAuthException exception = FirebaseAuthException(
-      message: details != null ? details['message'] : error['message'],
-      code: details != null ? details['code'] : 'unknown',
-    );
-
-    callbacks.verificationFailed(exception);
-  }
-
-  Future<void> _handlePhoneCodeSent(
-    Map<dynamic, dynamic> arguments,
-  ) async {
-    final int handle = arguments['handle'];
-    final String verificationId = arguments['verificationId'];
-    final int? forceResendingToken = arguments['forceResendingToken'];
-
-    PhoneAuthCallbacks callbacks = _phoneAuthCallbacks[handle]!;
-    callbacks.codeSent(verificationId, forceResendingToken);
-  }
-
-  Future<void> _handlePhoneCodeAutoRetrievalTimeout(
-    Map<dynamic, dynamic> arguments,
-  ) async {
-    final int handle = arguments['handle'];
-    final String verificationId = arguments['verificationId'];
-
-    PhoneAuthCallbacks callbacks = _phoneAuthCallbacks[handle]!;
-    callbacks.codeAutoRetrievalTimeout(verificationId);
   }
 
   /// Gets a [FirebaseAuthPlatform] with specific arguments such as a different
@@ -360,19 +281,23 @@ class MethodChannelFirebaseAuth extends FirebaseAuthPlatform {
   @override
   Stream<UserPlatform?> authStateChanges() async* {
     yield currentUser;
-    yield* _authStateChangesListeners[app.name]!.stream;
+    yield* _authStateChangesListeners[app.name]!
+        .stream
+        .map((event) => event.value);
   }
 
   @override
   Stream<UserPlatform?> idTokenChanges() async* {
     yield currentUser;
-    yield* _idTokenChangesListeners[app.name]!.stream;
+    yield* _idTokenChangesListeners[app.name]!
+        .stream
+        .map((event) => event.value);
   }
 
   @override
   Stream<UserPlatform?> userChanges() async* {
     yield currentUser;
-    yield* _userChangesListeners[app.name]!.stream;
+    yield* _userChangesListeners[app.name]!.stream.map((event) => event.value);
   }
 
   @override
@@ -612,23 +537,59 @@ class MethodChannelFirebaseAuth extends FirebaseAuthPlatform {
       );
     }
 
-    int handle = MethodChannelFirebaseAuth.nextMethodChannelHandleId;
-
-    _phoneAuthCallbacks[handle] = PhoneAuthCallbacks(verificationCompleted,
-        verificationFailed, codeSent, codeAutoRetrievalTimeout);
-
     try {
-      await channel
-          .invokeMethod<void>('Auth#verifyPhoneNumber', <String, dynamic>{
+      final eventChannelName = await channel
+          .invokeMethod<String>('Auth#verifyPhoneNumber', <String, dynamic>{
         'appName': app.name,
-        'handle': handle,
         'phoneNumber': phoneNumber,
         'timeout': timeout.inMilliseconds,
         'forceResendingToken': forceResendingToken,
         'autoRetrievedSmsCodeForTesting': autoRetrievedSmsCodeForTesting,
       });
+
+      EventChannel(eventChannelName!)
+          .receiveBroadcastStream()
+          .listen((arguments) {
+        final name = arguments['name'];
+        if (name == 'Auth#phoneVerificationCompleted') {
+          final int token = arguments['token'];
+          final String? smsCode = arguments['smsCode'];
+
+          PhoneAuthCredential phoneAuthCredential =
+              PhoneAuthProvider.credentialFromToken(token, smsCode: smsCode);
+          verificationCompleted(phoneAuthCredential);
+        } else if (name == 'Auth#phoneVerificationFailed') {
+          final Map<dynamic, dynamic>? error = arguments['error'];
+          final Map<dynamic, dynamic>? details = error?['details'];
+
+          FirebaseAuthException exception = FirebaseAuthException(
+            message: details != null ? details['message'] : error?['message'],
+            code: details != null ? details['code'] : 'unknown',
+          );
+
+          verificationFailed(exception);
+        } else if (name == 'Auth#phoneCodeSent') {
+          final String verificationId = arguments['verificationId'];
+          final int forceResendingToken = arguments['forceResendingToken'];
+
+          codeSent(verificationId, forceResendingToken);
+        } else if (name == 'Auth#phoneCodeAutoRetrievalTimeout') {
+          final String verificationId = arguments['verificationId'];
+
+          codeAutoRetrievalTimeout(verificationId);
+        }
+      });
     } catch (e) {
       throw convertPlatformException(e);
     }
   }
+}
+
+/// Simple helper class to make nullable values transferable through StreamControllers.
+class _ValueWrapper<T> {
+  const _ValueWrapper(this.value);
+
+  const _ValueWrapper.absent() : value = null;
+
+  final T? value;
 }