+++ packages/firebase_auth/firebase_auth_platform_interface/test/platform_interface_tests/platform_interface_auth_test.dart
@@ -35,8 +35,14 @@ void main() {
       );
       handleMethodCall((call) async {
         switch (call.method) {
-          case 'Auth#registerChangeListeners':
-            return {};
+          case 'Auth#registerIdTokenListener':
+            const String name = 'idTokenChannel';
+            handleEventChannel(name);
+            return name;
+          case 'Auth#registerAuthStateListener':
+            const String name = 'authStateChannel';
+            handleEventChannel(name);
+            return name;
           default:
             return null;
         }