+++ packages/firebase_performance/test/firebase_performance_test.dart
@@ -362,6 +362,20 @@ void main() {
         ]);
       });
 
+      test('requestPayloadSizeNullValue', () {
+        testMetric.requestPayloadSize = null;
+
+        expect(httpMetricLog, <Matcher>[
+          isMethodCall(
+            'HttpMetric#requestPayloadSize',
+            arguments: <String, dynamic>{
+              'handle': currentTestMetricHandle,
+              'requestPayloadSize': null,
+            },
+          ),
+        ]);
+      });
+
       test('responseContentType', () {
         testMetric.responseContentType = 'hi';
 
@@ -385,6 +399,20 @@ void main() {
             arguments: <String, Object?>{
               'handle': currentTestMetricHandle,
               'responsePayloadSize': 12,
+            },
+          ),
+        ]);
+      });
+
+      test('responsePayloadSizeNullValue', () {
+        testMetric.responsePayloadSize = null;
+
+        expect(httpMetricLog, <Matcher>[
+          isMethodCall(
+            'HttpMetric#responsePayloadSize',
+            arguments: <String, dynamic>{
+              'handle': currentTestMetricHandle,
+              'responsePayloadSize': null,
             },
           ),
         ]);